src/main/java/com/automationexercise/Pages/TestCasesPage.java
package com.automationexercise.Pages;

import org.openqa.selenium.By;
import org.openqa.selenium.WebDriver;
import org.openqa.selenium.WebElement;
import org.openqa.selenium.support.ui.ExpectedConditions;
import org.openqa.selenium.support.ui.WebDriverWait;

public class TestCasesPage extends BasePage {

    public TestCasesPage(WebDriver driver) {
        super(driver);
    }
    private By verification_ele =  By.xpath("//*[@id=\"form\"]/div/div[2]/h5/span");
    public boolean isTestCasesPageVisible() {
        WebElement element = driver.findElement(verification_ele);
        WebDriverWait wait = new WebDriverWait(driver, java.time.Duration.ofSeconds(10));
        wait.until(ExpectedConditions.visibilityOf(element));
        return element.isDisplayed();
    }
}

